+++ src/main/kotlin/klib/extensions/Char.kt
@@ -1,0 +1,11 @@
+package klib.extensions
+
+/**
+ * Converts a Char into a Number (Not Ascii)
+ *
+ * @return Char as Number
+ *
+ * @since 2.0.0
+ * @author Thomas Obernosterer
+ */
+fun Char.toStrInt() = this.toString().toInt()